+++ src/test/java/hu/nero/SubwayTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.*;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
